simplified IPC protocol by using one list atom
instead of sending argument by argument. Now only a UTF8 List Atom is
used to store the arguments of the command to call.

WARNING: IPC protocol has been changed and is incompatible to previous
versions. (So old clients cannot talk to a new server and a new client
cannot talk to old server)