completion: added parameter_expected system
With this, it can be found out, if a command expects a parameter. Only
if the command expects a parameter, then a completion for the parameter
will be searched. A typical example for this is a command like:

    floating [TAG] on|off|toggle|status

So the second parameter only is completable if the first one is a valid
tag. With the g_parameter_expected-table it can be found out if there is
a parameter for the current arguments.