ewmh: removing support for _NET_WORKAREA
This reverts commit 62d59d62802f2aae326085c328b1f50fe49e1e40. (and also
removes some lines of NEWS)

The support for the _NET_WORKAREA property has been removed due to
strange problems with menus of Qt4 applications: The menu didn't appear
or appeared at a wrong position with a wrong geometry.

The only problem is that EWMH specifies the _NET_WORKAREA property as a
mandatory property. Removing this property causes herbstluftwm to be not
EWMH-compliant anymore.

(and by the way the _NET_WORKAREA definition itself is in conflict with
the tag concept of herbstluftwm)