use XGetTextProperty for strings of unknown length
Do not use XGetWindowProperty() because it needs many calls until a
string of a (at compile time) unknown length is received. This causes a
race-condition if the value is changed between calls of
XGetWindowProperty() and can cause invalid values or even a xerror.

Example: Say you have a property with the value "123456789" and a
bufsize of 4.  Then the first XGetWindowProperty call returns "1234". If
the value is changed to "123" and XGetWindowProperty is called again
with offset 4 then it would call a xerror. If the value is changed to
"abcdefgh" then the result would be "1234efgh".

Many thanks to Florian Bruhin (The-Compiler) for spotting this bug and
for making it reproducible!