fix window focusing on window remove
The commit 840c3202b899d3dc829697a8528f1704ca878c48 did it right if the
currently selected window is removed but did it wrong in all other cases
(because the selection index always was decreased by 1, i.e. the window
focus cycled by -1 each time a window from this frame was removed).

Finally the behaviour is as desired:

Now the following behaviour is more intuitive: Say you have multiple
windows in one frame. If one window appears and the user closes it
immediately, then the original window is focused again.