Use CC from the environment for compiling
The Makefile now uses ${CC} defined in the environment if it's defined,
and only falls back to gcc otherwise.

Note: ${LD} is not used, but instead initialized to $(CC) in the
Makefile, since it's rare that one wants to use a separate linker these
days and it's also very likely to break.