Default to 0 for unspecified rectangle components
When parsing rectangle-specifications of the form widthxheight+x+y,
set missing components to zero.

This is done, using the return value of XParseGeometry, which tells us
what values it was able to parse. Of course, this is still not proper
error handling. Width and height probably shouldn't be allowed to be
omitted. Also, XParseGeometry doesn't report actual parsing errors.