Duplicate command args on keybinding execution
Ensure that the command args of a keybinding are not freed during the
execution of that command.

Actually this is not needed until now, because currently command args
only can be freed by keyunbind which does not access the command args
after freeing them.

It will be needed in the near future when chaining of commands is
possible. I.e. the chain exists of: 1. Remove the keybinding, 2. Do
something else.