ewmh: Register a new client after setting its tag
This solves this former lost-update race condition: Consider a panel
getting the event that there is a new client: The panel will read the
unset tag property of a new spawned client. Then hlwm sets the tag
property just before the panel starts listening for property changes on
the new client. => The panel won't notice what's the actual tag of that
client.

Thanks to Maximilian Krüger and Tobias Werth.