Quick fixup: Ensure shift_edge terminates
Prevent shift_edge from an endless loop in the following situation.
Consider a tag with (at least) two windows with the IDs 0xA and 0xB and
with the following frame layout and settings:

  herbstclient load '(clients horizontal:0 0xA 0xB)'
  herbstclient set focus_follows_shift 0
  herbstclient set default_direction_external_only 0

Then this command triggers an endless loop:

  herbstclient shift_edge right

This causes shift_edge to constantly shifting the window in that frame
to the right, which just swaps the windows. So there again is a window
at index 0 which can be shifted again. I.e. the layout cycles infinetly
between these two layouts:

  - (clients horizontal:0 0xA 0xB)
  - (clients horizontal:0 0xB 0xA)

The quick fix is two remember focus_follows_shift, set it to 1 and reset
it to its original state after the loop. A more clean solution should be
found soon.