Lock monitors during the handling of each X event
This locks all monitors for each X event and unlocks them again after
the event handling is finished. This prevents monitor_apply_layout() to
be called multiple times, which caused flickering for complex
operations.

Example: Frame A on monitor 0, Frame B and C on monitor 1, where B was
focused last. Each frame contains one window. Monitor 0 is focused. When
the user now clicks on the window in frame C, then the focus first
switched to frame B and then to frame C. As the monitors now are locked,
the focus directly jumps to frame C.