Place attribute values at constant positions
Keep the values of user-defined and temporary attributes at a constant
position in memory so that their adress does not change if the adress of
the according HSAttribute changes (e.g. because it is realloced when
creating a new attribute).

This also names two unions HSAttributePointer and HSAttributeValue. The
last one only is allocated if it is a user-defined or temporary
attribute. This case is indicated by user_data not being NULL and
pointing to the value. In this case also the name has to be freed when
freeing the attribute.

Strings are now saved as GString* and not as a pointer to a GString*.