Fixup decoration scheme member order
In the static initializations in decoration.c of the scheme triples,
the border width is named first and then the color. This order makes
more sense and is easier to read for the initialization, so also use
this for compatibility in the order in the struct definition.

It is a pitty that the C type system did not protect us here from
assigning a HSColor (which is an unsigned long typedef) to an int.