Fix decoration flicker on client unmap
This fixes the bug that a client flickered in the decoration color on
unmap, e.g. when the client quits. It is solved by placing a nothing
doing window behind the client. So this background window does not paint
anything so the actual decoration is not painted there if the client
disappears. So there's no flickering.

The code is a little bit different from the main decoration influence
openbox. In the openbox code the bgwin's colormap for clients with argb
visual is the default colormap for the openbox instance but this somehow
causes a BadMatch in XCreateWindow -- at least if DefaultColormap() is
used. If we use the client's colormap it works fine.