Delete WM_STATE property on client unmanage
Delete the WM_STATE so clients know that they have been unmanaged and
may show up again. This fixes bugs for applications which showed the
same dialog window multiple times (e.g. the connection window of
qjackctl).

For the ICCCM specification see reparenting comments in the comments on
reparenting after the "Advice to Implementors" comment in the "Changing
Window State" section:

http://www.x.org/releases/X11R7.7/doc/xorg-docs/icccm/icccm.html#Changing_Window_State