Let WM_STATE property be a XA_CARDINAL
Although the ICCCM http://tronche.com/gui/x/icccm/sec-4.html#s-4.1.3.1
requests, that the WM_STATE property is of type WM_STATE, some
applications have problems when herbstluftwm sets it to be an WM_STATE.
Other WMs like pekwm and openbox set it to a WM_STATE and it seems to
work.

If we set it to type WM_STATE, then katarakt does not recognize its new
size after being shifted from a visible tag to an invisible workspace
(s.t. the size changes). After setting an XA_CARDINAL, the size update
works again.

It's still open, why just setting WM_STATE to be of type WM_STATE and
filling it with two values (the state and 0, just like pekwm/openbox)
does not work. So this is just a workaround until we find the actual
problem or until bigger problems appear.