introduce Root and HookManager
Root is a singleton object that holds the new object tree.
HookManager is a simple node at the first level that holds all hooks.
It is trivial to add or remove hooks through the respective Actions.
This means that the user can add and remove her needed Hooks by using
the object tree.
Single commit as HookManager needs the Root to add hooks.