Migrate GString* output to std::ostream
Change every command as follows:

-int close_command(int argc, char** argv, GString* output);
+int close_command(int argc, char** argv, Output output) {;

This means that we also have to replace all occourences of g_string_*()
on these output parameters by << notation. When calling such a function,
the GString* output had to be replaced by a std::ostringstream.

Furthermore:

  - To reduce redundancy, the << operator was overloaded for Rectangle.
  - At some point, we had to replace char* by std::string
  - For object children printing we need to pass a ostream-pointer,
    that's not nice but will be removed soon anyway.

Note that the old object system still uses GString*.

There are still some hacky parts, which are marked by STRTODO as a
comment.