Make shift action give more predictable behavior
When shifting a window into another container, it would add it after the active client.

This was a bit strange & not obvious where it would add,
as well as being different to the behavior of directional focus shifting.

This patch changes behavior to add to the start/end, based on the movement direction.