Update focus objects when removing a tag
It is not optimal that the tags.focus object gets removed and
instantly recreated afterwards. However, I do not think that it is worth
the effort to redesign it properly before it gets refactored anyway in
the winterbreeze branch.

Thanks to ctx for spotting and fixing this bug. This solves #58