Fix issue with commands having more completions
When trying to complete "herbstclient use <tab>" it would not provide correct results as the space before tab was not considered and it only provided completions for command starting with "use". Additionally it printed the string "herbstclient" which was due to a missing --quiet parameter when using fish builtin string.