Make position monitor-relative on configure request
A client's floating window position is relative to the monitor the
client is on. However, the coordinates of a configure request is
relative to the root window. So convert the coordinates. If the client
is not visible, we fuzzily find the most appropriate monitor based on
the coordinates of the client.