Reload g_tree_style from settings on change
The string g_tree_style has to be updated each time the setting changes
because the string is reallocated and possibly changes its address. This
was formerly done by fetch_frame_colors(), which was not the right place
to do it. Not knowing that, the fetch_frame_colors() callback was
removed in commit

  ba31d3a48a80a0f00bcffe06f02c5149f036eb07 (in pull request #63).

Now, tree_style is updated explicitly in utils.cpp.