Fix Xlib race conditions in main loop
The Issue
---------
In the presence of many herbstclient calls, it happened that select() in
the main loop to block. This issue can be triggered with a loop like:

  for i in {0..1000} ; do herbstclient echo $i ; done

But it also occurred during the autostart or in q3terminal.sh.

This issue seems to be caused only on recent setups and not before
February 2018.

Unblocking select()
-------------------
If hlwm freezes during the autostart, the select() unblocks on any
events from X, e.g. pressing an hlwm keybinding or moving the cursor
across a window edge.

The Solution
------------
Instead of using XPending() we now flush the buffers manually by XSync()
and then read the length of the event queue directly by XQLength(). It
seems to me, that the XSync within the while(XQLength()) { ... } fixes
this issue, whereas the other changes are just code simplification
without changing the program semantics.