Fix compiler selection .travis.yml; general rework (#151)
* Fix compiler selection .travis.yml; general rework

Setting CC and CXX as environment variables gets overwritten by Travis
before the "script" is executed. The previous configuration basically
always used the default compiler.

To fix this, the variables are now set in a single-line "before_install"
script.

This change prints the CC/CXX values right before calling make, to
allow for proper manual verification.

As it turned out, the compilers we intended to use were not even
available. Installation of the right compiler is fixed here, too. Along
the way, the range of compiler is extended.

* Quote echo string

* Remove colon in echo (confuses Travis...)

* Exclude cast-function-type from -Werror on gcc-8

* Use -Wno-error= instead of flat out disabling