Improve test suite output if hlwm crashes
Explicitly check for undesired termination of herbstluftwm in the
fixtures when a client action times out and print an appropriate error
message.

In the hlwm_process fixture, don't wait() for herbstluftwm if the process
has been wait()'ed already. If a client action times out, the herbstluftwm
process has already been wait()'ed and so we avoid a second unnecessary
failed assertion.