Fix python version selection .travis.yml (#219)
The "python" key is only meaningful if the selected language is
"python". Selecting "cpp" is of not much use to us since we explicitly
install a specific version anyway.

Because Python 3.7 is not available on 14.04, the version spec is moved
into the xenial-based job that executes the tests. Other jobs do not
care about Python.