Cleanup interface of Input (#221)
Input is an ArgList that also stores the name of the command that
was called. It can only be constructed from a command name and (optionally) a
list of arguments to prevent wrong use.

All accessors from ArgList can be used. operator>> is overridden to
return the correct reference.

Also adds missing tests for substitute and sprintf commands.

Note that the use of shared_ptr in Input is *required* for CommandBinding::commandFromCFunc() to work. It might be better to just bite the bullet and copy the argument array there.