Travis: Use xvfb differently on 16.04 and 14.04 (#283)
Apparently, a recent change on Travis' side means that /etc/init.d/xvfb
does not exist on 16.04 anymore. Using the "services" approach is a
viable workaround on 16.04 but does not work on 14.04.

Abandoning the common script definition (which was only used by 2 of 4
jobs anyway) allows us to solve this as needed on the individual machine
types.