Move ownership of keybindings list into KeyManager
For this, the KeyBinding class is moved into keymanager.h, but not yet
turned into a nested class under KeyManager, in order to minimize the
diff here. Changing the class location will be easier later on.

The main goal and benefit here is getting rid of key_init/key_destroy.