Rework key module to be new and shiny OOP
This introduces a new class XKeyGrabber that handles the actual
grabbing/ungrabbing of keys and encapsulates most of the "weird"
knowledge required for it (like the current numlock mask).

Some more functionality has moved into KeyManager:
 - keymask handling
 - removal command (keyunbind)

The Keymask class is turned into a nested class of KeyManager as it is
only used internally.

Little remains in key.cpp/key.h, which can be removed in subsequent
steps.