Clean up responsibilities of KeyCombo helpers
Dealing with the structure of a KeyCombo (i.e., last token is keysym,
others are modifiers) is now part of fromString(). This makes the
renamed helpers for parsing modifiers and keysyms more flexible for
other use cases (completion in particular).

The helper for splitting a string into tokens is made public for future
use in completion code in KeyManager.

As a small cleanup, a local variable in mouse.cpp is renamed from
"string" to "str" to avoid shadowing the type.