Fix trailing whitespace in tree printing (#396)
* Fix trailing whitespace in tree printing

This makes appendCaption() responsible for putting a space between the
tree characters and the node caption. That way, the space can be omitted
when there is no caption at all (e.g., for the root node).

As a side-benefit, the multi-line string defined for testing this now
is free of any trailing whitespace, which makes flake8 happy, too.

* Update frametree, too