Implement the load command (#388)
Implement the load command. Other than in the old implementation in
master, this implementation makes a clear distinction between parsing
and application. The new class FrameParser is only responsible for
parsing, and the resulting tree is then merged with the existing frame
tree in order to have the same behaviour as before.