Fix keyboard focus (#476)
* Fix keyboard focus

The focus setting helper frame_focus_recursive() was called in many
places (and the focus was still broken when the first client showed up).
However, every such call is followed by a call to Monitor::applyLayout()
which sets the object link clients.focus. If the window is not visible
yet, XSetInputFocus() has no effect. So we need to reset the focus on
MapNotify events.

Hence, we now directly set the keyboard focus at this single place in
Monitor::applyLayout().

* Whitespace fixes