Check in herbstclient if herbstluftwm is running (#480)
* Check in herbstclient if herbstluftwm is running

Before sending commands or listening for hooks, check that herbstluftwm
is running. This is done by reading the window id of herbstluftwm's hook
window and then checking that this window still exists.

In the test suite, it is checked that herbstclient recognizes an absent
herbstluftwm. Previously, sending commands without herbstluftwm made
herbstclient hang.

In this new test case, we either don't start hlwm or we make it quit
abnormally, so we don't want the hlwm_process fixture to start and check
that hlwm shuts down gently. Hence, this removes the autouse-flag from
the hlwm_process fixture. This does not affect the other taste cases,
because they actively rely on hlwm which depends on hlwm_process.

* fixtures: Use hlwm_spawner in hlwm_process

* Move assert to hlwm_spawner

* subprocess: use universal_newlines=True

* Replace second text= with universal_newlines=