Wrapper for original client list and XQueryTree()
Write XConnection::queryTree() and a function in Ewmh to access the
original value of _NET_CLIENT_LIST. This removes two global variables in
ewmh.cpp and two functions from utils.h

This simplifies the scan() function. This commit tries to make it
obvious that the logic of scan() is not changed by this commit.