Move layout and align names to framedata.cpp
Move the arrays containing the names of the layout algorithms and the
frame split types to framedata.cpp. Other commands are supposed to use
the Converter-instance in order to print or parse these names.

For compatibility with the layout code, there is a temporary duplicate
of these arrays in command.cpp which can be removed as soon as the
commands set_layout and cycle_layout are migrated to the new
command/completion api.