Focus dummy window if no client is focused (#477)
* Fix keyboard focus

The focus setting helper frame_focus_recursive() was called in many
places (and the focus was still broken when the first client showed up).
However, every such call is followed by a call to Monitor::applyLayout()
which sets the object link clients.focus. If the window is not visible
yet, XSetInputFocus() has no effect. So we need to reset the focus on
MapNotify events.

Hence, we now directly set the keyboard focus at this single place in
Monitor::applyLayout().

* Whitespace fixes

* Focus dummy window if no client is focused

Previously, if no window is focused, the root windows was focused. This
had the effect that the keyboard input was sent to the window under the
mouse cursor (even though this windows was not focused).

We now create a dummy window outside of the screen. This window is used
already for EWMH's _NET_SUPPORTING_WM_CHECK, and now this window is
focused if no real client is focused.

Additionally, this fixes the analogue issue with focus_follows_mouse
which did not work previously when not client window is focused.

* Remove debug output

* Remove import