Save previous tag when stealing from other monitor
When `use`ing a tag that was already shown on another monitor and
swap_monitors_to_get_tag was set, we did not save the current tag in
monitor->previous_tag which resulted in use_previous not jumping back to
the previous tag but the one before that.

This adapts 4cc3c2adef257675953500d03519fc9a5ecba087 and also does the
same thing for the 'other' monitor.