Remove window_set_visible(), inline event masks
Remove the expensive and actually unneeded window_set_visible() and do
XMapWindow() and XUnmapWindow() where we actually need it. Note that
window_set_visible() was so complex because it has been used for
clients, formerly.

This allows us to inline the definitions of ROOT_EVENT_MASK and
CLIENT_EVENT_MASK.