Encapsulate mouse drag handling to own class (#606)
Encapsulate all the details what to do with mouse movemeng in a new
class MouseDragHandler. The responsibilities are now:

  - MouseManager has to get the events and mouse coordinates from X and
    pass them over to MouseDragHandler.
  - MouseDragHandler has to interpret mouse coordinate updates
    appropriately.
  - MouseDragHandler has to check whether the resources that are
    manipulated still exist and throw an DragNotPossible exception
    otherwise (these 'resources' are essentially the member variables of
    MouseDragHandler).

With this interface defined, MouseManager does not need the 'mode'
member anymore because the mode is implicit in a single other member
variable dragHandler_.

The benefit is that we can now cleanly implement mouse dragging features
for the tiling mode that may require other resources.