Mouse resizing of frames (#619)
In tiling mode, allow the resizing of frames with the mouse. This
loosely is based on ed2cd33 that implements the same functionality in
the old master code base. Still, resizing of empty frames is not yet
possible.

This also disables the 'null-dereference' error check since it
generates false positives for me all the time.

This closes #520.