Preserve focus and client order after frame removal (#621)
When removing a frame, insert the clients of the removed frame in the
target frame at the beginning or the end, depending on how the clients
were aligned visually before the split. Moreover, preserve the client
focus.

In total, the new semantics of the 'remove' command is less 'removing
the frame' but more 'undoing the split'.