Apply rules on clients whose WM_CLASS changes
According to ICCCM, the WM_CLASS has to be set before a client files a
maprequest. Clients (e.g. spotify) that set the WM_CLASS later, lack
the correct WM_CLASS entry when the rules are evaluated. Thus, whenever
the WM_CLASS of an existing client changes, apply the rules again such
that the rules matching against the class name are taking into account.

This solves #643.