Fix rule consequence for invalid tag names
Do not crash for the rule tag=sometag when no tag named 'sometag'
exists. We fix this in two places: where the tag name is looked up and
in the EWMH function that updates the EWMH tag property of the window,
just to be sure.

Thanks to @fiete201 for finding and reporting this issue.