Do not ignore transient-for-windows on startup (#707)
This fixes a bug that exists since df7d2717, because I didn't understand
the dwm sourcecode back than. On startup, the window manager must also
manage the clients that already exist. Since dwm uses the transient-for
hint to place dialog windows smartly (namely on the same tag as their
master window), dwm first processes the non-transient-for windows and
then the transient-for windows afterwards.

Back then, I only had the second part as a comment and then removed it
in 92c54e1da. Leading to the bug that all transient-for-windows were
ignored on hlwm startup. (For example the 'About' dialog of thunar is a
transient-for-window)

the present commit fixes this bug and adds a test case to avoid this in
the future.