Introduce keys_inactive client attribute (#715)
The current winterbreeze implementation of keymask was wrong (negated).
Since this is proven to be useful, to block certain keybindings for
concrete clients, we call this feature 'keys_inactive'.

This makes 'keymask' currently an attribute without any effect. The next
step is then to bring back the old behaviour of 'keymask'

See also #713.