Reimplement keymask
Implement the keymask attribute (with the meaning as in 0.7.1 and in the
docs). Also add some test cases for it; they are essentially the negated
cases as for keys_inactive.

Also add some more test cases for keys_inactive to ensure that this
commit does not mess up the logic of that.