Fix active smart_window_surroundings (#722)
Handle smart_window_surroundings entirely in HSFrameLeaf::
computeLayout() and not (as formerly) in Client::resize_tiling(). Here,
we have all the necessary data available to decide whether the
window_gap needs to be applied or not.

Due to the migration, the relevant code in Client has been turned into a
commend, and with the rewrite/cleanup the client has no access to the
frame it lives in.

Also add a test case to check that the new computations are correct.

This solves issue #721. Thanks to @tiosgz for reporting this!