Separate cycle_all implementation from command (#752)
This is another step towards #732. Here, the implementation of the
cycle_all command is separated from the command, and even the wrapping
is now configurable. So now, it's easy to let cycle_all also traverse
the floating windows.

The test cases still fully cover the functionality as it is at the
moment.