Retry if the display closes during X startup
For some reason in the test suite, the display sometimes closes during
the initialization phase. My only explanation for this is that the
display still exists from a previous test case and that this display is
about to close.

One attempt is to simply retry opening the connection after two seconds.
It will take some time until we'll see whether this works around the
issue.