Never compile a regex for the empty string
An empty string is not allowed in the Posix grammar for extended regular
expressions. On OpenBSD[1], this led to a run time error on startup
because we initialize the keymask regexes with the empty string.

In the keymask code, we are using the empty string anyway as a special
case for an 'inactive' or 'unset' regular expression, so there is no
need to compile the regex for "".

Thanks to Lucas <lucas@sexy.is> and Florian Viehweger
<openbsd@out-of-creativity.de> for spotting and reporting this issue.

[1] OpenBSD clang version 8.0.1 (tags/RELEASE_801/final) (based on LLVM 8.0.1)