'resize' command: make delta parameter optional
Recently the 'resize' command became also functional in the floating
mode, where the 'FRACTIONDELTA' parameter is not used at all. In the
tiling mode, the parameter is most likely '0.02' for most people. Hence,
it makes sense to make the parameter optional, so one can now run e.g.

   herbstclient resize up

to make the focused client (or frame) grow into the upward direction.