Update fish completion script (#839)
* Only complete herbstclient options where appropriate

Herbstclient options must precede, if there is any, the first command to
herbstluftwm. Do not offer them as arguments to hlwm commands.

* Avoid spaces for partial completions

Fixes #797 with a workaround.

As fish does not seem to support partial completions, we detect partial
completions and eagerly fetch their subsequent completions.

This includes a switch to hlwm's complete_shell method:
 • Escape sequences are removed in COMMAND and ARGS.
 • Special characters will be escaped in the output.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>