ICCCM: Make WM_STATE of type WM_STATE
Make the WM_STATE property of every window have the type WM_STATE, i.e.
apply what is in a5045733. When I originally tried to adapt this commit
to the new codebase I accidentally changed the type back to XA_CARDINAL
(84a6ed87), thus effectively reverting the fix a5045733.

There's also a test case for this now to avoid that this bug is
introduced a third time.

This fixes issue #849.