close_(and/or)_remove only close in floating mode
For floating clients (regardless of single window or the entire tag),
the commands close_and_remove and close_or_remove do not make sense
because either the window is in no frame (single window floating) or
because the frame is not visisble (tag is floated). Hence these commands
must behave like the ordinary 'close' if the focused client is floated.

On the development side, this causes that we need to move both commands
from the FrameTree to the HSTag class. This commit also adds test cases
to verify both the new behavior and the usual behavior in tiling mode.

This fixes #853.