Enable theme setting tight_decoration again (#858)
The tight_decoration attribute specifies whether the window sizehints
affect the decoration window or only the window content itself. Other
than in the old code base of v0.7.2, we now always center the window if
the size hints change the window size.

This only affects tiled clients with sizehints_tiling enabled. A visual
demonstration of the tight_decoration setting can be found in the
pull-request for this commit:
https://github.com/herbstluftwm/herbstluftwm/pull/858

Effectively, this commit only removes a redundant "if (!false)" and
replaced one "if (false)" by "if (scheme.tight_decoration())". The rest
is documentation and tests.