Focus frames with the mouse
This adds basic support for selecting frames with the mouse: either by
click or by mouse hover (if focus_follows_mouse is enabled). This change
only concerns the actual frame. If there is no frame shown (because
always_show_frame=off) or if there is a hole in the frame (because
frame_bg_transparent=on), then we only obtain events for the root window
that are currently ignored. (Interpreting event coordinates on the root
window is in the scope of another change)

On the implementation level, this introduces a static map that
transforms window ids into frame decorations (I think a 'framemanager'
is overkill here). The focusFrame() function is added to the root
window, because it affects the global state of hlwm (and I think
focus_client() should go to Root at some point, too).

This solves issue #20 (in principle) and adds a test case for it.