Tests: Fix race condition in mouse button fixture (#890)
When waiting for hlwm to process the mouse button events, it does not
suffice to wait for the respective 'ButtonPress' or 'ButtonRelease'
debug output on stderr, because this only indicates that hlwm is
starting to process the respective mouse button events. So if we
continue doing something via our own X connection, then the hlwm might
not be finished processing the button event.

Hence, for wait=True, we run the 'true' command because as soon as hlwm
processes some command, we know that the event handling is finished.

We do something similar for mouse move events.

I suppose/hope that this fixes #889.