Fix EWMH urgency add for clients without wm hints (#892)
This fixes issue #586 where a client without wm hints (here, the
chromium window) is equipped with the urgency hint via ewmh, e.g. by:

    wmctrl -r Chromium -b add,demands_attention

In such a case, the urgency hint was not updated correctly, and the
tag_flags hook was not fired. This commit fixes both by creating a new
wm hint for clients lacking it instead of doing nothing. Also, a test
for this is added.