Update fish completions (#854)
* Only complete herbstclient options where appropriate

Herbstclient options must precede, if there is any, the first command to
herbstluftwm. Do not offer them as arguments to hlwm commands.

* Avoid spaces for partial completions

Fixes #797 with a workaround.

As fish does not seem to support partial completions, we detect partial
completions and eagerly fetch their subsequent completions.

This includes a switch to hlwm's complete_shell method:
 • Escape sequences are removed in COMMAND and ARGS.
 • Special characters will be escaped in the output.

* Only consider relevant part of command line

A regression introduced in 73c64e78d0 would lead to false completions
when hc is completed in a chain or pipe (e.g. echo hi; and hc <tab>)

* Correctly cut off preceding herbstclient options

Note: that test always passed, and was also misguided

* Do not complete hlwm commands where not expected

* Ignore commands after the cursor

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>