Tests: Fix race condition in test_reload (#904)
The issue was that the call() command processed the output of the
herbstluftwm process and so the context manager shutdown code could not
find the strings it was looking for.

There was essentially the same issue in test_spawn that never appeared
in the CI runs (i.e. "in the wild").

The present commit mainly fixes the race condition. While at it, I also
fixed a minor (maybe only theoretic) race condition in
read_and_echo_output() itself, where a line on stdout/stderr was lost if
it is not enclosed by a '\n' byte within the timeout.

This fixes #902.