Don't drop later EnterNotify during enternotify()
This fixes a race condition, reported in #908.

In general, Monitor::applyLayout() possibly drops future EnterNotify
events in the X11 event queue that might have been artifically triggered
because the new tiling on a monitor might have moved a window to the
cursor.

During the handling of an EnterNotify event, no such artificial enter
notify events can be created, because only the focus is changed, so
there is not necessity to drop those. Moreover, on quick mouse
movements, multiple enter notify events can be created if the cursor
traverses over multiple windows. In this case, we must not drop these
later EnterNotify events.

(These events can be triggered while hlwm is already processing the
first EnterNotify event, so it is not enough to skip to the newest
EnterNotify event in the enternotify() event handler).

To accomplish this restriction, the dropEnterNotifyEvents() function is
moved to XMainLoop where we keep track of whether we are in
enternotify(). On a positive side, we get rid of calling an X function
in monitor.cpp.

There are two new test cases:

  - The first test case verifies that after resizing, non-welcome
    EnterNotify events are dropped.
  - The second one checks that for a number of EnterNotify events, the
    last one survives. Unfortunately, I did not manage to establish a
    test case that fails before the present commit. (But I think this is
    better than nothing because it still tests something)

In order to test this manually, do the following:

  - herbstclient set focus_follows_mouse on
  - herbstclient set frame_border_width 20
  - put two clients in separate frames side by side
  - move the mouse cursor quickly between the two clients.
  - before the present commit, the focus isn't updated reliably, but
    with this commit, it is. (Instead of switching between git commits
    one can also simply deactivate the 'return' in
    XMainLoop::dropEnterNotifyEvents())