Create Xvfb manually in the test cases
This introduces a new fixture 'xvfb' that creates and Xvfb instance
whose scope is only the test case and not the entire test session. The
python code for this has already been written in the MultiscreenDisplay
class.

Having the Xvfb scope only within the testcase gets rid of the necessity
to kill existing/remaining windows at the beginning and the end of the
testcase run.

Moreover, I am passing -noreset to Xvfb, and so the X properties are not
reset when the last client window closes, hence we don't need this
x11_connection fixture hack anymore.