Quickfix race-condition in hc --wait testcases (#964)
There is a race-condition in both test cases for herbstclient --wait
that comes from the fact that we don't know how fast herbstclient
connects to hlwm and starts listening for hooks. In this commit, we
simply hope that `hc --wait` boots up within the time needed for two `hc
true` calls.

This issue only exists for --wait and not for --idle, because for
--wait, we need to test that herbstclient indeed exits after a fixed
number of hooks (--count).

For hc --idle, this issue does not appear in the HcIdle class in
conftest.py, because there, we can wait for hc --idle to boot up by
calling 'emit_hook' multiple times (hence, this is not applicable for
synchronizing the bootup of --wait).