Handle changes of the tag attr 'floating_focused'
When changing the attribute 'floating_focused' of a tag the tag needs a
relayout. This commit connects the signals accordingly. We don't want the
unecessary relayout when the change of 'floating_focused' comes from
hlwm itself, e.g. within commands like 'jumpto' or 'cycle_all'. Hence,
this commit introduces a new signal on attributes that is only fired
when the use triggers a change, more specifically, when
Attribute::change() is called. The new testcase verifies this commit
(i.e. the testcase failed before this commit).

This fixes #947.

While implementing this, I noticed a bug in the Signal-types and the
overloading of Signal::connect(). The non-working call of connect() is
added as a comment, but fixing this issue is beyond the scope of this
commit.